---  base_decorator.rb
@@ -1,13 +1,0 @@
-require_relative 'nameable'
-
-class Decorator < Nameable
-    def initialize(nameable)
-        super()
-        @nameable = nameable
-    end
-
-    def correct_name
-        @nameable.correct_name
-    end
-
-end
