+++ capitalize_decorator.rb
