+++ person.rb
@@ -5,6 +5,8 @@ class Person < Nameable
   attr_accessor :name, :age
 
   def initialize(age, name = 'Unknown', parent_permission: true)
+    super(age, parent_permission)
+
     @id = rand(1..1000)
     @name = name
     @age = age
